+++ src/lib.rs
@@ -1,3 +1,5 @@
+pub mod validate;
+
 pub fn hello() {
     println!("Hello from semver!");
 }