+++ HISTORY.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+### Added
+
+- Added `hello` subcommand
+- Added `validate` subcommand that checks whether the provided version complies to semver
+
 ## [v0.0.0 - 2021-11-13](https://github.com/se7entyse7en/semver/compare/486f8cd34136f830e21c15ff179a74a251165fd9...v0.0.0)
 
 - Project inception! :tada: